pyproject.toml
[project]
name = "az-reporter"
version = "0.1.0"
description = "Azure Resource Graph reporter (DevOps interview tool)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "azure-identity>=1.17.0",
  "backoff>=2.2.1",
  "pydantic>=2.6.0",
  "requests>=2.32.0",
  "typer>=0.12.0",
]

[project.scripts]
azr = "az_reporter.cli:main"
